Implement player emote system and refactor item management
- Introduced PlayerEmoteManager to handle emote display upon failed apple consumption.
- Updated CheckIfPlayerHasApples to reference GameController for apple tracking.
- Refactored ItemManager to manage apple quantities through GameController.
- Enhanced ShortcutManager with new skill shortcuts for improved gameplay interaction.
- Updated PlayerActions input actions to include new skill shortcuts.
- Adjusted PlayerStatsManager and HUDManager to support float values for health and spirit points.
- Added emote-fail sprite asset for visual feedback during gameplay.

+++ Assets/Scripts/CharactersData/PlayerStatsData.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 public class PlayerStatsData : CharacterStatsData
 {
     [Header("Current Stats")]
-    public int _currentHP;
-    public int _currentSP;
+    public float _currentHP;
+    public float _currentSP;
 
     public void ResetStats()
     {
@@ -13,26 +13,26 @@ public class PlayerStatsData : CharacterStatsData
         _currentSP = MaxSP;
     }
 
-    public int GetCurrentHP()
+    public float GetCurrentHP()
     {
         return _currentHP;
     }
 
-    public int GetCurrentSP()
+    public float GetCurrentSP()
     {
         return _currentSP;
     }
 
     public void LoadFromPrefs()
     {
-        _currentHP = PlayerPrefs.GetInt("PlayerHP", _currentHP);
-        _currentSP = PlayerPrefs.GetInt("PlayerSP", _currentSP);
+        _currentHP = PlayerPrefs.GetFloat("PlayerHP", _currentHP);
+        _currentSP = PlayerPrefs.GetFloat("PlayerSP", _currentSP);
     }
 
     public void SaveToPrefs()
     {
-        PlayerPrefs.SetInt("PlayerHP", _currentHP);
-        PlayerPrefs.SetInt("PlayerSP", _currentSP);
+        PlayerPrefs.SetFloat("PlayerHP", _currentHP);
+        PlayerPrefs.SetFloat("PlayerSP", _currentSP);
         PlayerPrefs.Save();
     }
 }
